Revert "Improvements to DataTable types and fixed vulnerabilities (#294)"
