+++ src/DataTable.tsx
@@ -6,6 +6,7 @@
  */
 
 import React, { useCallback, useEffect, useRef } from 'react';
+import PropTypes from 'prop-types';
 import {
   useSortBy,
   usePagination,
@@ -15,10 +16,13 @@ import {
   useRowSelect,
   useFilters,
   useGlobalFilter,
+  Filters,
+  FilterValue,
   useColumnOrder,
   Row,
 } from 'react-table';
 import classNames from 'classnames';
+import { DataTableColumn, DataTableColumnInstance } from './types';
 import {
   DATATABLE_DEFAULT_PAGE_SIZES,
   checkBoxSelectionHook,
@@ -31,206 +35,197 @@ import DataTableHeaderCell from './DataTableHeaderCell';
 import DataTableDragdropProvider from './DataTableDragdropProvider';
 import DataTableContextMenuProvider from './DataTableContextMenuProvider';
 import useDataTableInstance from './useDataTableInstance';
-import { DataTableProps, DataTablePropTypes } from './DataTable.props';
 
-function RenderFilterPanel<T extends Record<string, unknown>>(
-  props: DataTableProps<T>,
-  tableInstance: any,
-) {
-  const { filterPanel, disableFiltering } = props;
-  const {
-    filterColumns,
-    resetFilter,
-    resetAllFilters,
-    setGlobalFilter,
-    state: { filters: activeFilters, globalFilter },
-  } = tableInstance;
-
-  return filterPanel && !disableFiltering
-    ? filterPanel({
-        filterColumns,
-        activeFilters,
-        resetFilter,
-        resetAllFilters,
-        globalFilter,
-        setGlobalFilter,
-      })
-    : null;
+export interface DataTableProps<T extends Record<string, unknown>>
+  extends Omit<React.HTMLProps<HTMLDivElement>, 'data' | 'size' | 'as'> {
+  id: string;
+  data: ReadonlyArray<T>;
+  columns: ReadonlyArray<DataTableColumn<T>>;
+  striped?: boolean;
+  bordered?: boolean;
+  borderless?: boolean;
+  hover?: boolean;
+  size?: string;
+  variant?: string;
+  responsive?: boolean | string;
+  pageSize?: number;
+  pageSizeOptions?: number[];
+  resizeColumns?: boolean;
+  multipleRowSelection?: boolean;
+  checkBoxRowSelection?: boolean;
+  disableRowSelectionOnClick?: boolean;
+  disablePagination?: boolean;
+  disableSorting?: boolean;
+  disableFiltering?: boolean;
+  disableDragging?: boolean;
+  stickyFirstColumn?: boolean;
+  filterPanel?: (
+    columns: DataTableColumnInstance<T>[],
+    filters: Filters<T>,
+    resetFilter: (columnId: string) => void,
+    resetAllFilters: () => void,
+    globalFilter: any,
+    setGlobalFilter: (filterValue: FilterValue) => void,
+  ) => React.ReactElement;
+  dragTemplate?: (column: DataTableColumnInstance<T>) => React.ReactElement;
+  onRowSelectionChange?: (rows: any[]) => void;
 }
 
-function RenderPagination<T extends Record<string, unknown>>(
-  props: DataTableProps<T>,
-  tableInstance: any,
-) {
-  const { data, disablePagination, pageSizeOptions, size } = props;
-  const {
-    gotoPage,
-    setPageSize,
-    state: { pageIndex, pageSize },
-  } = tableInstance;
+const propTypes = {
+  /** A HTML id attribute, necessary for proper form accessibility. */
+  id: PropTypes.string.isRequired,
 
-  return !disablePagination ? (
-    <TablePagination
-      count={data.length}
-      pageIndex={pageIndex}
-      pageSize={pageSize}
-      onPageChange={gotoPage}
-      pageSizeOptions={pageSizeOptions || DATATABLE_DEFAULT_PAGE_SIZES}
-      onPageSizeChange={setPageSize}
-      size={size}
-    />
-  ) : null;
-}
+  /**
+   * An array of column data of type `DataTableColumn` for the Table Headers.
+   */
+  columns: PropTypes.array.isRequired,
 
-function RenderTableHeader<T extends Record<string, unknown>>(
-  props: DataTableProps<T>,
-  tableInstance: any,
-) {
-  const { size, dragTemplate } = props;
+  /**
+   * Default page size if pagination is enabled.
+   */
+  pageSize: PropTypes.number,
 
-  const {
-    headerGroups,
-    allColumns,
-    visibleColumns,
-    getAllHeadersInAGroup,
-    toggleHideColumn,
-    toggleHideAllColumns,
-    setColumnOrder,
-  } = tableInstance;
+  /**
+   * An array of page size options if pagination is enabled.
+   */
+  pageSizeOptions: PropTypes.arrayOf(PropTypes.number),
 
-  return (
-    <thead className="bg-gray-light">
-      {headerGroups.map((headerGroup) => (
-        <tr {...headerGroup.getHeaderGroupProps()} role="row">
-          <DataTableContextMenuProvider
-            size={size}
-            allColumns={allColumns}
-            toggleHideAllColumns={toggleHideAllColumns}
-            toggleHideColumn={toggleHideColumn}
-          >
-            <DataTableDragdropProvider
-              visibleColumns={visibleColumns}
-              setColumnOrder={setColumnOrder}
-              dragItemTemplate={dragTemplate}
-            >
-              {getAllHeadersInAGroup(headerGroup.headers, headerGroup.id).map(
-                (column, index: number) => (
-                  <DataTableHeaderCell
-                    key={column.id}
-                    header={column}
-                    onToggleHideColumn={toggleHideColumn}
-                    role="columnheader"
-                    aria-sort={
-                      (column.isSorted &&
-                        (column.isSortedDesc ? 'descending' : 'ascending')) ||
-                      'none'
-                    }
-                    className="bg-gray-light sticky-top"
-                    aria-colindex={index + 1}
-                  >
-                    {column.render('Header')}
-                  </DataTableHeaderCell>
-                ),
-              )}
-            </DataTableDragdropProvider>
-          </DataTableContextMenuProvider>
-        </tr>
-      ))}
-    </thead>
-  );
-}
+  /**
+   * An array of data for the Table Rows.
+   */
+  data: PropTypes.array.isRequired,
 
-function RenderTableBody<T extends Record<string, unknown>>(
-  props: DataTableProps<T>,
-  tableInstance: any,
-  handleRowClickFn: (row: Row<T>) => void,
-) {
-  const { checkBoxRowSelection, disablePagination } = props;
+  /**
+   * Enables column resizing.
+   */
+  resizeColumns: PropTypes.bool,
 
-  const { rows, page, prepareRow, getTableBodyProps } = tableInstance;
+  /**
+   * Enables checkbox to select the rows.
+   */
+  checkBoxRowSelection: PropTypes.bool,
 
-  return (
-    <tbody {...getTableBodyProps()}>
-      {(disablePagination ? rows : page).map((row) => {
-        prepareRow(row);
-        return (
-          <tr
-            {...row.getRowProps()}
-            onClick={() => handleRowClickFn(row)}
-            className={classNames(row.isSelected && 'selected')}
-            role="row"
-          >
-            {row.cells.map((cell, index: number) => (
-              <td
-                {...cell.getCellProps(getCellStyles)}
-                className={classNames(
-                  index === 0 &&
-                    `${
-                      checkBoxRowSelection && 'icon-only checkbox-selector-cell'
-                    }`,
-                )}
-                role="cell"
-                aria-rowindex={index + 1}
-              >
-                {cell.render('Cell')}
-              </td>
-            ))}
-          </tr>
-        );
-      })}
-    </tbody>
-  );
-}
+  /**
+   * Enables multiple row selection.
+   */
+  multipleRowSelection: PropTypes.bool,
 
-function RenderTable<T extends Record<string, unknown>>(
-  props: DataTableProps<T>,
-  tableInstance: any,
-  handleRowClickFn: (row: Row<T>) => void,
-) {
-  const {
-    data,
-    striped,
-    stickyFirstColumn,
-    bordered,
-    borderless,
-    hover,
-    size,
-    variant,
-    responsive,
-  } = props;
+  /**
+   * Disables on-click row selection.
+   */
+  disableRowSelectionOnClick: PropTypes.bool,
 
-  const { allColumns, isResizing, getTableProps } = tableInstance;
+  /**
+   * Displays all the rows and hides pagination panel.
+   */
+  disablePagination: PropTypes.bool,
 
-  return (
-    <Table
-      striped={striped}
-      bordered={bordered}
-      borderless={borderless}
-      hover={hover}
-      size={size}
-      variant={variant}
-      responsive={responsive}
-      {...getTableProps()}
-      role="table"
-      aria-colcount={allColumns.length}
-      aria-rowcount={data.length}
-      className={classNames(
-        stickyFirstColumn && 'table-sticky-first-column',
-        isResizing && 'resizing',
-      )}
-    >
-      {RenderTableHeader(props, tableInstance)}
-      {RenderTableBody(props, tableInstance, handleRowClickFn)}
-    </Table>
-  );
-}
+  /**
+   * Removes sorting icons from the column header.
+   */
+  disableSorting: PropTypes.bool,
+
+  /**
+   * Removes filter panel.
+   */
+  disableFiltering: PropTypes.bool,
+
+  /**
+   * Disables dragging function on Table headers.
+   */
+  disableDragging: PropTypes.bool,
+
+  /**
+   * Adds zebra-striping to any table row within the `<tbody>`.
+   */
+  striped: PropTypes.bool,
+
+  /**
+   * Adds borders on all sides of the table and cells.
+   */
+  bordered: PropTypes.bool,
+
+  /**
+   * Removes all borders on the table and cells, including table header.
+   */
+  borderless: PropTypes.bool,
+
+  /**
+   * Enable a hover state on table rows within a `<tbody>`.
+   */
+  hover: PropTypes.bool,
+
+  /**
+   * Make tables more compact by cutting cell padding in half by setting
+   * size as `sm`.
+   */
+  size: PropTypes.string,
+
+  /**
+   * Pins the first column.
+   */
+  stickyFirstColumn: PropTypes.bool,
+
+  /**
+   * Invert the colors of the table — with light text on dark backgrounds
+   * by setting variant as `dark`.
+   */
+  variant: PropTypes.string,
+
+  /**
+   * Responsive tables allow tables to be scrolled horizontally with ease.
+   * Across every breakpoint, use `responsive` for horizontally
+   * scrolling tables. Responsive tables are wrapped automatically in a `div`.
+   * Use `responsive="sm"`, `responsive="md"`, `responsive="lg"`, or
+   * `responsive="xl"` as needed to create responsive tables up to
+   * a particular breakpoint. From that breakpoint and up, the table will
+   * behave normally and not scroll horizontally.
+   */
+  responsive: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),
+
+  /**
+   * Callback when rows selected or unselected.
+   *
+   * ```js
+   * function onRowSelectionChange (rows: any[]) => void
+   * ```
+   */
+  onRowSelectionChange: PropTypes.func,
+
+  /**
+   * Function that returns a custom filter panel.
+   *
+   * ```js
+   * function filterPanel(columns: DataTableColumnInstance<T>[], filters: Filters<T>, resetFilter: (columnId: string) => void,
+   *            resetAllFilters: () => void, globalFilter: any, setGlobalFilter: (filterValue: FilterValue) => void
+   *            ) => React.ReactElement
+   * params:
+   *  columns - An array of filter columns, refer `DataTableColumnInstance`
+   *  activeFilters - An array of active filters
+   *  resetFilter - Function to reset a column's filter
+   *  resetAllFilters - Function to reset all the filters
+   *  globalFilter - Global filter value
+   *  setGlobalFilter - Function to set the global filter value
+   * ```
+   */
+  filterPanel: PropTypes.func,
+
+  /**
+   * Function that returns a template for Dragging item.
+   *
+   * ```js
+   * function dragTemplate(column: DataTableColumnInstance<T>) => React.ReactElement
+   *  column - An instance of the column object that is dragged
+   * ```
+   */
+  dragTemplate: PropTypes.func,
+};
 
 function DataTable<T extends Record<string, unknown>>(
   props: React.PropsWithChildren<DataTableProps<T>> & {
     ref?: React.Ref<HTMLDivElement>;
   },
 ): React.ReactElement {
-  /* eslint-disable */
   const {
     id,
     columns,
@@ -258,17 +253,16 @@ function DataTable<T extends Record<string, unknown>>(
     className,
     filterPanel,
     dragTemplate,
-    ...htmlProps
+    ...rest
   } = props;
-  /* eslint-enable */
-
   const resolvedRef = (useRef<HTMLDivElement>(null) ||
     ref) as React.MutableRefObject<HTMLDivElement>;
 
   const enableRowSelection =
     !disableRowSelectionOnClick || checkBoxRowSelection;
-  const disableMouseClickRowSelection =
-    disableRowSelectionOnClick || checkBoxRowSelection;
+  const enableRowSelectionOnClick = !(
+    disableRowSelectionOnClick || checkBoxRowSelection
+  );
 
   // Conditional Table Hooks Array for react-table
   const conditionalHooks: any = [useFlexLayout];
@@ -297,13 +291,29 @@ function DataTable<T extends Record<string, unknown>>(
     ...(!multipleRowSelection && stateReducer),
   };
 
-  const tableInstance = useDataTableInstance<T>(
-    columns,
-    data,
-    tableOptions,
-    conditionalHooks,
-  );
-  const { selectedFlatRows } = tableInstance;
+  const {
+    headerGroups,
+    rows,
+    page,
+    allColumns,
+    selectedFlatRows,
+    filterColumns,
+    visibleColumns,
+    isResizing,
+    state: { pageIndex, pageSize, filters, globalFilter },
+    prepareRow,
+    getTableProps,
+    getTableBodyProps,
+    gotoPage,
+    setPageSize,
+    getAllHeadersInAGroup,
+    resetFilter,
+    resetAllFilters,
+    setGlobalFilter,
+    toggleHideColumn,
+    toggleHideAllColumns,
+    setColumnOrder,
+  } = useDataTableInstance<T>(columns, data, tableOptions, conditionalHooks);
 
   // Callback APIs
   useEffect(() => {
@@ -313,11 +323,11 @@ function DataTable<T extends Record<string, unknown>>(
 
   const handleRowClick = useCallback(
     (row: Row<T>) => {
-      if (!disableMouseClickRowSelection) {
+      if (enableRowSelectionOnClick) {
         row.toggleRowSelected(!row.isSelected);
       }
     },
-    [disableMouseClickRowSelection],
+    [enableRowSelectionOnClick],
   );
 
   return (
@@ -328,10 +338,19 @@ function DataTable<T extends Record<string, unknown>>(
         bordered && 'mrb-data-table-bordered',
         className,
       )}
-      {...htmlProps}
+      {...rest}
     >
       <div className={classNames('d-flex flex-column overflow-hidden')}>
-        {RenderFilterPanel(props, tableInstance)}
+        {filterPanel &&
+          !disableFiltering &&
+          filterPanel(
+            filterColumns,
+            filters,
+            resetFilter,
+            resetAllFilters,
+            globalFilter,
+            setGlobalFilter,
+          )}
         <div
           className={classNames(
             'mrb-data-table-container d-flex flex-column overflow-hidden',
@@ -340,16 +359,115 @@ function DataTable<T extends Record<string, unknown>>(
           {
             // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
             <div className={classNames('scrollable')} tabIndex={0}>
-              {RenderTable(props, tableInstance, handleRowClick)}
+              <Table
+                striped={striped}
+                bordered={bordered}
+                borderless={borderless}
+                hover={hover}
+                size={size}
+                variant={variant}
+                responsive={responsive}
+                {...getTableProps()}
+                role="table"
+                aria-colcount={allColumns.length}
+                aria-rowcount={data.length}
+                className={classNames(
+                  stickyFirstColumn && 'table-sticky-first-column',
+                  isResizing && 'resizing',
+                )}
+              >
+                <thead className="bg-gray-light">
+                  {headerGroups.map((headerGroup) => (
+                    <tr {...headerGroup.getHeaderGroupProps()} role="row">
+                      <DataTableContextMenuProvider
+                        size={size}
+                        allColumns={allColumns}
+                        toggleHideAllColumns={toggleHideAllColumns}
+                        toggleHideColumn={toggleHideColumn}
+                      >
+                        <DataTableDragdropProvider
+                          visibleColumns={visibleColumns}
+                          setColumnOrder={setColumnOrder}
+                          dragItemTemplate={dragTemplate}
+                        >
+                          {getAllHeadersInAGroup(
+                            headerGroup.headers,
+                            headerGroup.id,
+                          ).map((column, index: number) => (
+                            <DataTableHeaderCell
+                              key={column.id}
+                              header={column}
+                              onToggleHideColumn={toggleHideColumn}
+                              role="columnheader"
+                              aria-sort={
+                                (column.isSorted &&
+                                  (column.isSortedDesc
+                                    ? 'descending'
+                                    : 'ascending')) ||
+                                'none'
+                              }
+                              className="bg-gray-light sticky-top"
+                              aria-colindex={index + 1}
+                            >
+                              {column.render('Header')}
+                            </DataTableHeaderCell>
+                          ))}
+                        </DataTableDragdropProvider>
+                      </DataTableContextMenuProvider>
+                    </tr>
+                  ))}
+                </thead>
+                <tbody {...getTableBodyProps()}>
+                  {(disablePagination ? rows : page).map((row) => {
+                    prepareRow(row);
+                    return (
+                      <tr
+                        {...row.getRowProps()}
+                        onClick={() => handleRowClick(row)}
+                        className={classNames(row.isSelected && 'selected')}
+                        role="row"
+                      >
+                        {row.cells.map((cell, index: number) => (
+                          <td
+                            {...cell.getCellProps(getCellStyles)}
+                            className={classNames(
+                              index === 0 &&
+                                `${
+                                  checkBoxRowSelection &&
+                                  'icon-only checkbox-selector-cell'
+                                }`,
+                            )}
+                            role="cell"
+                            aria-rowindex={index + 1}
+                          >
+                            {cell.render('Cell')}
+                          </td>
+                        ))}
+                      </tr>
+                    );
+                  })}
+                </tbody>
+              </Table>
             </div>
           }
-          {RenderPagination(props, tableInstance)}
+
+          {!disablePagination && (
+            <TablePagination
+              count={data.length}
+              pageIndex={pageIndex}
+              pageSize={pageSize}
+              onPageChange={gotoPage}
+              pageSizeOptions={pageSizeOptions || DATATABLE_DEFAULT_PAGE_SIZES}
+              onPageSizeChange={setPageSize}
+              size={size}
+            />
+          )}
         </div>
       </div>
     </div>
   );
 }
 
-DataTable.propTypes = DataTablePropTypes;
+DataTable.propTypes = propTypes;
 
 export default DataTable;
