+++ src/react-table-config.d.ts
@@ -112,8 +112,6 @@ declare module "react-table" {
   }
 }
 
-
-
 // export type TableMouseEventHandler = (instance: TableInstance<T>) => MouseEventHandler
 
 export interface UseFlexLayoutColumnPropsCustom<D extends object = {}> {
